ml_backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pymongo>=4.6.0
cloudinary>=1.41.0
python-dotenv>=1.0.1
requests>=2.31.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
opencv-python>=4.9.0.80
easyocr>=1.7.1
deepface>=0.0.93
pytesseract>=0.3.10
mediapipe>=0.10.8
Pillow>=10.3.0
# Optional for blockchain integration later
web3>=6.18.0
